requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.93.2
numpy==2.2.6
pandas==2.3.1
requests==2.32.4
fastapi-cors==0.0.6
mangum==0.17.0